Implement sanitizer builds. (#54)
This fixes #9.  The environment variables
SANITIZE_ADDRESS, SANITIZE_MEMORY, SANITIZE_UNDEFINED,
SANITIZE_THREAD, and SANITIZE_LEAKS can be set in the
CI builds to enable the corresponding sanitizer.

The build script examines the output of the tests searching
for sanitizer errors, and the CI build fails if any is found.

Developers can use:

```console
$ CXX=clang++ CC=clang cmake -DSANITIZE_ADDRESS=ON
```

to manually enable the sanitizer in their build.