Upgrade to gRPC v1.10.x (#283)
Fix #230.  Update gRPC to v1.10.x.  There are known bugs in
v1.9.x that manifest when using round robin load balancing
and disconnects.

We want to enable -Wall -Werror on our code, but we do not want
to force it on the user, nor do we want to force it on our
dependencies.  I needed to make the change now to workaround the
build problems referenced in:

https://github.com/grpc/grpc/issues/14511

But they are good changes in general too.

Use `cmake -H. -B.build` to simplify script.