Cleanup CMAKE_FLAGS hack. (#1078)
I was abusing CMAKE_FLAGS to pass information down to the external
projects. Instead we should rely (when needed), on environment variables
if the platform requires special settings. Standard variables, like
BUILD_SHARED_LIBS, should be passed down using -DVAR=${VAR}.

Disable shared libraries for Fedora build.

I was trying to be clever, but we cannot (yet) test shared libraries
with projects that do not install the dependencies. We are using
external projects, which install the shared libraries in non-standard
locations, and the binaries cannot find them.