Workaround Travis CI build time problems. (#1801)
Until we have a better solution in place this enables us to bootstrap
the build cache. Basically when starting with a completely empty cache
the build will stop halfway, which will partially populate the cache and
make the next build go faster.

The majority of this change is to ensure that "halfway" captures enough
things to make for a substantial initial cache.

The algorithm to decide if the cache is "cold" is kind of naive, we
really test for "the cache has 0 bytes". Obviously this will not work
sometimes, for example, if the cache is full of garbage. In that case we
can manually clear the cache in Travis and start over.

I created #1800 to track a real solution, because these hacks are not
awesome.