Automate testing of INSTALL.md instructions. (#1991)
With this change the INSTALL.md file is automatically generated from the
several `ci/test-readme/Dockerfile.*` files. In turn the Dockerfiles are
changed to test:

- That the minimal set of tools recommended in README.md are enough to
  compile `google-cloud-cpp` using external projects.

- That the instructions in the INSTALL.md file are enough to compile and
  install `google-cloud-cpp`.

- And that the installed `google-cloud-cpp` can be used to compile a
  sample program that a user might write.

A future PR (or series of them) should execute these Dockerfiles as part
of the CI build. In a future PR we should also extend the Dockerfiles to
execute one of the integration tests against the production environment,
to verify that the produced binaries are usable.

The post-install tests were hard-coded to use g++. On some platforms
(notably ubuntu:18.04) g++ compiles with PIC, and links with PIE, while
clang++ does not. This fixes the build to use the same compiler for the
post-install tests.