chore: Remove last Travis build (#2852)
Make the clang-tidy build perform all the same checks as the last
Travis build. In addition, this build now runs the same checks as build
with the same name in the Cloud Spanner project.

This is the last build on Travis, so I also removed the Travis
configuration file. A future PR (or probably more than one) will remove
the now unused Travis files.