cleanup: Use only find_package to find dependencies. (#2967)
This PR changes the CMake files to use `find_package()` exclusively when
looking for dependencies. To automatically download the dependencies the
project can be embedded in a larger "super build".

There are still some places where we use `find_package()` through a helper
file, because different versions of the CMake module (or config file) behave
differently and we want to manage those differences in one place.