chore: upgrade most builds to gRPC-1.24.3 (#3217)
Upgrade the Bazel and CMake super builds to gRPC-1.24.3. Did not update
all the builds because that requires more changes (upgrading Protobuf
for example). This fixes the build problems on openSUSE/Tumbleweed.

I had to make a code change because, depending on the Protobuf version,
the `GetReflection()` function generated in a proto may be static or
not. clang-tidy generates a warning when the function is static and we
use it as a non-static member function. The code does not compile when
used as a static member function.