chore: update g-c-cpp-common to v0.19.0 (#3384)
Upgrade to the latest version of `google-cloud-cpp-common`. This version
removed a number of symbols from `google::cloud::grpc_utils::internal`,
and we unfortunately used those. Better to bite the bullet and use the
new symbols everywhere, as opposed to the backwards compatibility
aliases.

In addition, the behavior of a test class (`EnvironmentVariableRestore`)
changed for the better, and some tests needed fixing.

Updated the API baseline because of the symbol name changes.

Fixed the Windows+CMake build scripts to correctly update vcpkg
dependencies.