ci: enable integration tests for asan+msan builds (#3839)
This PR enables the integration tests on the AddressSanitizer and
MemorySanitizer builds. The tsan and ubsan builds need more work before
we can enable them.

I had to change the `curl.BUILD` file: as it was the integration tests could
not run on Fedora, `libcurl` needs to know where the certificate authority
bundles are located. The CMake and Makefile builds for it figure this out
automatically, but we use Bazel+Fedora for the sanitizer builds, and the
`curl.BUILD` file had the CA bundle location hard-coded for Ubuntu.