fix: moved .dockerignore to ci/ (#3848)
* fix: moved .dockerignore to ci/

During the docker build phase, the `.dockerignore` file is searched for
in the "context", which in the case of our builds is the `ci/`
directory. So, I believe, previously the `.dockerignore` file was not
being used. Moving it into `ci/` seems to cause it to be used now in my
tests.

The result of this is that our docker images should be rebuilt less
often.

* removed ci/benchmarks, which wasn't used