chore: upgrade libcurl to v7.69.1 (#3851)
We have not updated this in a while, and there are important bug fixes
in newer versions. I would prefer not to chase these bugs if they
manifest in our builds. Note that in some platforms we build against the
version installed on the platform, which can be much older.

One of the test had hard-coded assumptions about the number of file
descriptors created by a `storage::Client`. The number depends on the
version of libcurl used under the hood. Where possible I just changed
the test to make no such assumptions, we are looking for file descriptor
leaks in this test, we do not really care how many are created.