fix: libcurl Bazel rules (#4461)
We have custom Bazel rules for libcurl, as the library does not support
Bazel. Our rules were definining CURL_MAX_WRITE_SIZE to 64KiB when
*compiling* libcurl, but used the default (16KiB) when *using* libcurl.
This macro defines the maximum buffer passed to the write callback, and
we use it in the storage library to size our buffers. You can now see
how this is a problem. Lucky for us, when running against production the
actual data received is around 1500 bytes (the MTU maybe?) so things
worked, but trying to run the tests against the emulator did not work.