support!: drop support for gcc 4.8 (#4796)
Part of: #4788

BREAKING CHANGE: This PR raises our oldest supported GCC version from 4.8 (the default installed on centos-7) to GCC 5.4, which is the oldest gcc that's installed on the Linux distros that are supported by GCE and that we test with. Ubuntu Xenian (16.04) comes with gcc 5.4.

This PR adds a new build named "gcc-5.4". In follow-up changes, I'll enable that build in our CI, and disable the existing "gcc-4.8" build.

This PR also changes our centos:7 "install" test to use gcc-7 (instead of 4.8) using the devtoolset-7 Software Collection.