chore(deps): upgrade abseil to 20210324.2 (#6664)
Fixes: #6627

This version of Abseil adds proper pkgconfig support, so we no longer
need dynamically compute the Abseil .so files to link with, instead we
can just list the pkg-config deps that we "require" from Abseil. Indeed,
our old way of computing the libraries no longer works, so this change
was required in the same PR as the Abseil upgrade.