ci: enable CMake policies up to 3.24 (#9912)
The main motivation is to improve the customer experience with CMake
under Windows.  Until CMake 3.15 switching the MSVC runtime was a bit of
a pain.  We used `cmake_minimum_required(VERSION 3.10)` which has two
effects:

- It requires CMake to be >= 3.10
- It guarantees the CMake 3.10 behavior, even if you use 3.15 or 3.24

We do not want the 3.10 behavior on Windows, it requires too many steps
to switch the MSVC runtime library.