emulator tests: Fix bigtable_emulator_range_set_test.
The tests aborted due to an assertion in the constructor of Range that
enforced the invariant that start <= end for all well-formed
Ranges. However in several tests testing the ends of ranges, start was
passed in as "unimportant" with end being "A" or "B". This violates
the invariant and causes an abort at runtime.

Fixes: TBL-62.