emulator: server: implement MutateRows.
This is a partial fix for TBL-60. It partially fixes that issue
because the test now successfully executes the bulk mutation that was
failing (it called the MutateRows RPC).
    
The test (bigtable_table_sample_rows_integration_test) still fails,
 but that happens at a subsequent location and that will be fixed in an
 upcoming PR.
    
This is presented as its own PR since it is a single, logical, useful change.
    
TESTED=That part of the integration test now passes.
    
References: TBL-60.