fix: emulator: Filters: Fix the ColumnRange Filter
This fixes the ColumnRange filter bug in which the filter would let through cells from all other column families.

The failing bigtable_filters_integration_test now passes and there is a unit test (now fixed in this PR) that checks for this specific bug and which is now passing.

An end-to-end test - in which we store data in an actual emulator table in multiple column families and apply the filter - is also now implemented in this PR and passing.

References: TBL-58
Fixes: TBL-58